creating test dir
