+++ CMakeLists.txt
@@ -3,4 +3,5 @@ cmake_minimum_required(VERSION 3.19)
 set(CMAKE_CXX_STANDARD 20)
 
 add_subdirectory(apps)
-#add_subdirectory(lib)
+add_subdirectory(lib)
+add_subdirectory(test)
